Crate tree_sitter_uvl

This crate provides uvl language support for the tree-sitter parsing library.

Typically, you will use the language function to add this language to a tree-sitter Parser, and then use the parser to parse some code:

let code = "";
let mut parser = tree_sitter::Parser::new();
parser.set_language(tree_sitter_uvl::language()).expect("Error loading uvl grammar");
let tree = parser.parse(code, None).unwrap();

Constants

NODE_TYPES

The content of the node-types.json file for this grammar.

Functions

language

Get the tree-sitter Language for this grammar.